backend/wakeword/detection/emergency.py
# detection/emergency.py
from channels.layers import get_channel_layer
from asgiref.sync import async_to_sync

class EmergencySystem:
    def __init__(self):
        self.channel_layer = get_channel_layer()
        
    def send_alert(self, message):
        async_to_sync(self.channel_layer.group_send)(
            "emergency_responses",
            {
                "type": "emergency.message",
                "content": message
            }
        )

    def handle_panic(self, confidence):
        """Execute emergency protocols"""
        if confidence > 0.8:
            self.trigger_sirens()
            self.notify_authorities()
            
    def trigger_sirens(self):
        # IoT integration
        pass
        
    def notify_authorities(self):
        # SMS/Email integrations
        pass