requirements.txt
django
channels
channels-redis
onnxruntime
torch
torchaudio
torchvision
twilio
librosa
numpy
scipy
datasets
audiomentations
torch-audiomentations
onnx2tf
onnx_tf
python-dotenv
tqdm
webrtcvad
djangorestframework
soundfile
speechbrain
pronouncing
tensorflow-cpu
protobuf
onnx
torchinfo
torchmetrics
tensorflow_probability
ai_edge_litert
onnx_graphsurgeon
sng4onnx
deep-phonemizer
mutagen
numba
acoustics
psutil
redis
python-rtmixer
webvtt-py
